refactor: improve idempotency for package_managers

+++ dotfiles/src/common/packages/package_managers.py
@@ -1,6 +1,7 @@
 import shutil
 
 from common.helpers import run_command
+from common.logger import logger
 from common.packages.models import Package, PackageManager
 
 
@@ -22,7 +23,8 @@ class Brew(PackageManager):
 
     @classmethod
     def update_all(cls) -> None:
-        run_command([cls.COMMAND, "upgrade"])
+        run_command([cls.COMMAND, "update"])  # Syncs formulae
+        run_command([cls.COMMAND, "upgrade"])  # Upgrades packages
 
 
 class Yay(PackageManager):
@@ -31,7 +33,9 @@ class Yay(PackageManager):
 
     @classmethod
     def install(cls, package: Package) -> None:
-        run_command([cls.COMMAND, "--sync", "--noconfirm", package.identifier])
+        run_command(
+            [cls.COMMAND, "--sync", "--needed", "--noconfirm", package.identifier]
+        )
 
     @classmethod
     def uninstall(cls, package: Package) -> None:
@@ -57,7 +61,7 @@ class Chocolatey(PackageManager):
 
     @classmethod
     def install(cls, package: Package) -> None:
-        run_command([cls.COMMAND, "install", package.identifier])
+        run_command([cls.COMMAND, "upgrade", package.identifier])
 
     @classmethod
     def uninstall(cls, package: Package) -> None:
@@ -65,7 +69,7 @@ class Chocolatey(PackageManager):
 
     @classmethod
     def update_all(cls) -> None:
-        run_command([cls.COMMAND, "upgrade", "all"])
+        run_command([cls.COMMAND, "upgrade", "all", "--confirm"])
 
 
 class Git(PackageManager):
@@ -74,14 +78,31 @@ class Git(PackageManager):
 
     @classmethod
     def install(cls, package: Package) -> None:
-        cmd = [cls.COMMAND, "clone", package.identifier]
-        if package.resolved_path:
-            cmd.append(str(package.resolved_path))
-        run_command(cmd)
+        if not package.resolved_path:
+            logger.error(
+                f"Package {package.identifier} requires a 'destination' for Git "
+                "operations."
+            )
+            return
+
+        if package.resolved_path.exists() and (package.resolved_path / ".git").exists():
+            # Pull changes for "update" logic
+            run_command([cls.COMMAND, "-C", str(package.resolved_path), "pull"])
+        else:
+            # Fresh clone
+            run_command(
+                [
+                    cls.COMMAND,
+                    "clone",
+                    "--recurse-submodules",
+                    package.identifier,
+                    str(package.resolved_path),
+                ]
+            )
 
     @classmethod
     def uninstall(cls, package: Package) -> None:
-        if package.resolved_path:
+        if package.resolved_path and package.resolved_path.exists():
             shutil.rmtree(package.resolved_path)
 
     @classmethod
